+++ borgweb/borg/models/cache.py
@@ -2,9 +2,9 @@ from django.db import models
 
 
 class Cache(models.Model):
-    total_chunks = models.IntegerField()
-    total_csize = models.IntegerField()
-    total_size = models.IntegerField()
-    total_unique_chunks = models.IntegerField()
-    unique_csize = models.IntegerField()
-    unique_size = models.IntegerField()
+    total_chunks = models.BigIntegerField()
+    total_csize = models.BigIntegerField()
+    total_size = models.BigIntegerField()
+    total_unique_chunks = models.BigIntegerField()
+    unique_csize = models.BigIntegerField()
+    unique_size = models.BigIntegerField()
